+++ tests/integration/cases/call_deep_dotted_transformed/Perl_call.pl
@@ -1,3 +1,7 @@
+sub app {}
+sub client {}
+sub fetch {}
+sub emit {}
 emit(app.client.fetch("hello"));
 emit(app.client.fetch(42));
 emit(app.client.fetch(1));