+++ tests/integration/cases/call_per_element_false/Perl_call.pl
@@ -1,1 +1,2 @@
+sub process {}
 process(1);